// sourcecode/BTrak.Packages/snova-hrmangement/module/module-components/src/lib/hrmanagment/models/employee-contact-details-model.ts
export class EmployeeContactDetailsModel{
    employeeContactDetailId: string;
    employeeId: string;
    stateId: string;
    address1: string;
    address2: string;
    postalCode: string;
    countryId: string;
    homeTelephone: string;
    mobile: string;
    workTelephone: string;
    workEmail: string;
    otherEmail: string;
    contactPersonName: string;
    relationshipId: string;
    dateOfBirth: Date;
    employeeContactTypeId: string;
    isArchived: boolean;
    userId: string;

    firstName: string;
    surName: string;
    userName: string;
    email: string;
    countryName: string;
    countryCode: string;
    timeStamp: any;
    totalCount: number;
    stateName: string;
}